Renamed project and did code refactoring from pysubnetter to pynetcal and PySubnetter to PyNetcal
