+++ .gitignore
@@ -1,5 +1,6 @@
 /dist/
 /pysubnetter.egg-info/
+/pynetcal.egg-info/
 /build/
 **/__pycache__/**
 **/.pytest_cache/**
