+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "python.pythonPath": "/usr/bin/python3",
     "python.testing.pytestArgs": [
-        "pysubnetter"
+        "pynetcal"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.nosetestsEnabled": false,
