+++ pynetcal/pynetcal.py
@@ -1,10 +1,10 @@
 from ipaddress import IPv4Address, IPv4Network
-import pysubnetter.ipv4networkclasses as ipv4helper
-from pysubnetter.ipv4subnet import IPv4Subnet
-from pysubnetter.ipv4subnetlist import IPv4SubnetList
-from pysubnetter.ipv4subnetmask import IPv4SubnetMask
+import pynetcal.ipv4networkclasses as ipv4helper
+from pynetcal.ipv4subnet import IPv4Subnet
+from pynetcal.ipv4subnetlist import IPv4SubnetList
+from pynetcal.ipv4subnetmask import IPv4SubnetMask
 
-class PySubnetter:
+class PyNetcal:
 
     @classmethod
     def ipv4_calculate_subnets_flsm(self, ipv4network, hosts=None,
