--- pysubnetter/__init__.py
@@ -1,1 +1,0 @@
-__package__ = "pysubnetter"
